fixed rendering of polygons from TIGER data.
polygons were, in the past, represented by a series of points.  the
points were the endpoints of the lines making up the polygon.  what
was missing were all the points in the middle of those lines, i.e.,
the "shape points".

the polygon table now holds the lines that make up the polygon, and
when traversing, the lines are looked up, and both the endpoints and
the shapepoints are rendered.

there's an "endianness" to lines in polygons -- you have to traverse
each line segment in the right order.  lines which should be traversed
in the "unnatural" order (i.e., "to" to "from") are flagged in the
polygon table by being stored as negatives.  the endpoints and shapepoints
of lines so flagged are reversed during rendering.

nothing was done to ensure correctness of "holes" (e.g. islands), or
of overlapping polygons (e.g. lakes within parks).